Move jetstream into its own proper test
